added some generic tests
