--- src/components/__test__/Whatever.test.js
@@ -1,9 +1,0 @@
-describe('true is truthy and false is falsy', () => {
-  test('true is truthy', () => {
-    expect(true).toBe(true);
-  });
-
-  test('false is falsy', () => {
-    expect(false).toBe(false);
-  });
-});
